requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
argcomplete==1.12.0
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
click==7.1.2
colorlog==4.2.1
coverage==5.2.1
distlib==0.3.1
Django==3.0.3
django-bootstrap4==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
freezegun==0.3.15
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
more-itertools==8.4.0
multidict==4.7.6
nox==2020.5.24
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-vcr==1.0.2
python-dateutil==2.8.1
pytz==2020.1
psycopg2==2.8.5
PyYAML==5.3.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
starlette==0.13.4
typing-extensions==3.7.4.2
uvicorn==0.11.3
uvloop==0.14.0
vcrpy==4.0.2
virtualenv==20.0.28
wcwidth==0.2.5
WebOb==1.8.6
websockets==8.1
wrapt==1.12.1
WSGIProxy2==0.4.6
yarl==1.5.0
zipp==3.1.0
